第九次更新：
完成所以MapReduce

Top250 ： 本项目亮点

+++ src/main/java/utils/HbaseUtils.java
@@ -38,14 +38,14 @@ public class HbaseUtils {
         admin.createTable(tdb.build());
     }
 
-    public static boolean isExists(Admin admin,String tableName) throws IOException {
+    public static boolean isExists(Admin admin, String tableName) throws IOException {
 
         return admin.tableExists(TableName.valueOf(tableName));
     }
 
-    public static void dropTable(Admin admin,String tableName) throws IOException {
+    public static void dropTable(Admin admin, String tableName) throws IOException {
 
-        if (isExists(admin,tableName)) {
+        if (isExists(admin, tableName)) {
 
             admin.disableTable(TableName.valueOf(tableName));
 
@@ -56,32 +56,37 @@ public class HbaseUtils {
         }
     }
 
-    public static void jobSubmission(Admin admin,String sourceTable,String targetTable,
-                                     Class<?> Map, Class<?> Reduce, Class<?> outputKey, Class<?> outputValue) throws IOException, ClassNotFoundException, InterruptedException {
+    public static boolean jobSubmission(Admin admin, String sourceTable, String targetTable,
+                                        Class<?> Map, Class<?> Reduce, Class<?> outputKey, Class<?> outputValue) throws IOException, ClassNotFoundException, InterruptedException {
 
         BasicConfigurator.configure();
-        if (isExists(admin,targetTable)) {
-            dropTable(admin,targetTable);
-        }
-        createTable(targetTable, new String[]{"Per_Info"},admin);
+        if (isExists(admin, sourceTable)) {
+            if (isExists(admin, targetTable)) {
+                dropTable(admin, targetTable);
+            }
+            createTable(targetTable, new String[]{"Per_Info"}, admin);
 
-        Job job = Job.getInstance(init());
-        job.setJarByClass(HbaseUtils.class);
-        Scan scan = new Scan();
-        scan.setCaching(100);
-        scan.setCacheBlocks(false);
-        TableMapReduceUtil.initTableMapperJob(
-                sourceTable,
-                scan,
-                (Class<? extends TableMapper>) Map,
-                (Class<? extends Writable>) outputKey,
-                (Class<? extends Writable>) outputValue,
-                job);
-        TableMapReduceUtil.initTableReducerJob(
-                targetTable,
-                (Class<? extends TableReducer>) Reduce,
-                job);
-        boolean b = job.waitForCompletion(true);
+            Job job = Job.getInstance(init());
+            job.setJarByClass(HbaseUtils.class);
+            Scan scan = new Scan();
+            scan.setCaching(100);
+            scan.setCacheBlocks(false);
+            TableMapReduceUtil.initTableMapperJob(
+                    sourceTable,
+                    scan,
+                    (Class<? extends TableMapper>) Map,
+                    (Class<? extends Writable>) outputKey,
+                    (Class<? extends Writable>) outputValue,
+                    job);
+            TableMapReduceUtil.initTableReducerJob(
+                    targetTable,
+                    (Class<? extends TableReducer>) Reduce,
+                    job);
+            boolean b = job.waitForCompletion(true);
+            return b;
+        } else {
+            return false;
+        }
     }
 
 }
